refactor: split scheduling helpers
